[Silver I] Title: 밤양갱, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/31926. 밤양갱/밤양갱.cc
@@ -19,10 +19,9 @@ void solution()
 {
     int ans = 10;
 
-    ll cri = 1;
+    int cri = 1;
     while (n >= cri * 2)
     {
-//        cout << "cri : " << cri << '\n';
         ans++;
         cri *= 2;
     }
